Remove Fixed Has Issues on Smaller Screens
With fixed the links are not visible only the header and image. Still in aside for now may change to bottom of report. For now it is near the beginning with icon (may remove icon later or just use icon to symbolize), Can now scroll on smaller screens to be able to see and click hyperlinks to the git repos